fix(geometry): split boolean/mod.rs trait wiring back under its ratchet budget
#3922 merged from a base 88 commits stale and grew
rust/geometry/src/processors/boolean/mod.rs to 948 lines against its
recorded budget of 936, turning main red on
rust/processing/tests/module_size_ratchet.rs. Every other gate was green;
only cargo test observes this one.

Move `impl GeometryProcessor for BooleanClippingProcessor` and `impl
Default` into a sibling `router_impl.rs`, leaving mod.rs at 921 lines. No
allowlist row is added and no budget is raised, per the ratchet's own
instruction to shrink or split.

Pure move: no geometry logic changes. `chain_cycle_tests.rs` reached the
trait through `use super::*`, so it now imports `GeometryProcessor`
directly rather than relying on a re-export it never named.
